app/pyproject.toml
[project]
name = "kiosco-cinemex"
version = "0.1.0"
description = "Backend FastAPI kiosco dulcería Cinemex"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "pillow>=11.0.0",
    "pywin32>=308; sys_platform == 'win32'",
    "qrcode[pil]>=8.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0"]

[tool.ruff]
line-length = 100
target-version = "py311"
